Feat: split BNF functions parse.c에 있던 함수들을 분리
