--- parser/tree.c
@@ -1,27 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   tree.c                                             :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: daeha <daeha@student.42.fr>                +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/05/27 16:14:58 by daeha             #+#    #+#             */
-/*   Updated: 2024/05/27 18:01:33 by daeha            ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "parser.h"
-
-t_node	*new_parent_node(t_node_type type, t_node *left, t_node *right)
-{
-	t_node	*node;
-
-	node = malloc(sizeof(t_node));
-	if (node == NULL)
-		//error here
-	node->cmd = NULL;
-	node->type = type;
-	node->left = left;
-	node->right = right;
-	return (node);
-}
